Adición: ArenitoPlugin
Un plugin para contener el código relativo al Arenito
